Fix fabricacion populate queries and add detail endpoint
